Task(2) - Random Password Generator
Ecstatic to share the completion of a significant internship task-2 ! The learning curve was steep, but the sense of accomplishment is unmatched. Can't wait to showcase the results and insights gained. Ready to take on more challenges ahead with this newfound experience! 

#oasisinfobyteinternship
#oasisinfobyte

#pythondevelopment

Task(2)-Random Password Generator
>>generate your password in three simple steps:-

>step1 : select the specific characters you want to include in your password.

>step2: specify the length of your password.

>step3 : select complexity of your password .
(passwords are directly copied to the clipboard once they are generated .)

Rules:
1-easy password only consists of alphabets and numbers.
2- normal password doesn’t contain any symbols.
3- specify length and complexity in order to generate passwords.
4-passwords are directly copied to the clipboard once they are generated.